+++ src/openassay/__init__.py
@@ -8,7 +8,9 @@ from openassay.acceptance import AcceptanceResult, LevelStats, run_acceptance
 from openassay.backcalc import BackCalcResult, Sample, back_calculate, back_calculate_many
 from openassay.curve import CalibrationResult, StandardCurve, fit_standard_curve
 from openassay.ingest import read_plate
+from openassay.parallelism import ParallelismResult, test_parallelism
 from openassay.plate import CollapsedReplicate, PlateData, PlateLayout, PlateWell, Well
+from openassay.potency import PotencyResult, relative_potency
 from openassay.range import RangeResult, determine_lloq_uloq
 from openassay.report import report_run
 
@@ -21,6 +23,8 @@ __all__ = [
     "PlateData",
     "PlateLayout",
     "PlateWell",
+    "ParallelismResult",
+    "PotencyResult",
     "CollapsedReplicate",
     "Sample",
     "StandardCurve",
@@ -29,7 +33,9 @@ __all__ = [
     "back_calculate_many",
     "fit_standard_curve",
     "determine_lloq_uloq",
+    "relative_potency",
     "read_plate",
     "report_run",
     "run_acceptance",
+    "test_parallelism",
 ]